Remove Bash support
I started using Bash in 1997. It was the only shell I used since then.
I created a utility (txt2regex) with Bash builtin commands only and
even wrote a 500-pages book on Bash programming in 2008.

I started using Fish in July 2020. After 23 years, trying out a
different shell was quite an experience. I forced myself to use
Fish-only since the start, to really commit and see if I would like it.
Well, I never came back.

I still kept updating both Fish and Bash configuration in this
repository, because "what if". It was never needed. It was just a waste
of time since I moved to Fish in all the systems I use.

In this commit I'm removing all the Bash-related configuration. The last
Bash state (the parent for this commit) is saved as the `bash` branch
and my plan is just leaving it there.

It's a bit sad, and a bit of a relief.